The hud text can now be optionally opaque
The fps counter and the texts at the bottom of the screens could sometimes be
unreadable when the game uses clear background, so this option allows to draw
the text on a black background. This is slightly slower than drawing
transparent text (+ it uses C code and the transparent text uses asm code), so
the default remains to use the transparent version.

Also makes the get_white_pen function reliable even in depths > 8bpp.

+++ source/video/palette.h
@@ -137,7 +137,7 @@ Used by the ingame gui, to get the white pen (nearly always 255)
 
 */
 
-UINT8 get_white_pen(void);
+UINT32 get_white_pen(void);
 
 // Call this to remove resources (eg. in clear default)
